requirements.txt
asgiref==3.8.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
configparser==6.0.1
Django==5.1.3
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-db-connection-pool==1.2.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
EditorConfig==0.17.0
et-xmlfile==1.1.0
exrex==0.11.0
Faker==37.1.0
ghp-import==2.1.0
greenlet==3.1.1
griffe==1.5.7
gunicorn==23.0.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.5
jsbeautifier==1.15.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.5
mkdocs-material-extensions==1.3.1
mkdocs-mermaid2-plugin==1.2.1
mkdocstrings==0.28.2
mkdocstrings-python==1.16.2
norminette==3.3.51
numpy==1.26.4
openpyxl==3.0.10
packaging==24.2
paginate==0.5.7
pandas==2.2.1
pathspec==0.12.1
pi==0.1.2
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.3
pync==2.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyyaml_env_tag==0.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-futures==1.0.1
rpds-py==0.22.3
setuptools==78.1.0
six==1.17.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparams==6.1.0
sqlparse==0.5.2
stem==1.8.2
torrequest==0.1.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.2
watchdog==6.0.0
